Add initial book edition and fix timezones.
It's still missing a lot of features. A issue where the subtitle was not
being saved was fixed on the server.

JOOQ seems to be converting any provided UTC date to the local timezone,
so dates in DTOs needs to be transformed to UTC during the model
conversion. The book DTOs are fixed on this matter, but the others still
are missing a fix.