fix(server): make book code non unique in database
The schema was incorrectly created with a unique constraint in the code field of Book, making it
impossible to create books with the same code in different libraries (even from different users).
Since it's not currently possible to remove constraints in SQLite databases, the server will restart
the database from scratch by using a different file name in the resources files. If the user has
customized the resource, it will need to delete or rename the database file, otherwise it will face
this issue while creating the same book in different libraries.

BREAKING CHANGE: The database file will be restarted from scrach since it's not possible to delete a
constraint in SQLite yet. The resources files were updated to force a new database to be created.
Data will be lost, but it's expected to these things to happens as it's a WIP app without a stable
release yet.